refactor: Attach main package to root folder
